requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
async-timeout==4.0.2
attrs==21.4.0
charset-normalizer==2.0.12
click==8.0.4
Flask==2.0.3
Flask-Seeder==1.2.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
frozenlist==1.3.0
greenlet==1.1.2
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
multidict==6.0.2
pkg_resources==0.0.0
psycopg2-binary==2.9.3
SQLAlchemy==1.4.31
typing_extensions==4.1.1
Werkzeug==2.0.3
WTForms==3.0.1
yarl==1.7.2
